pyproject.toml
[build-system]
requires = ["uv_build>=0.8.18"]
build-backend = "uv_build"

[project]
name = "agent-core"
version = "0.1.0"
description = "可擴充的 AI Agent 核心框架"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "anthropic>=0.76.0",
    "redis>=7.1.0",
]

[project.optional-dependencies]
app = [
    "fastapi>=0.128.0",
    "python-dotenv>=1.0.0",
    "uvicorn[standard]>=0.40.0",
]
mcp = ["mcp>=1.0.0"]

[dependency-groups]
dev = [
    "fastapi>=0.128.0",
    "python-dotenv>=1.0.0",
    "uvicorn[standard]>=0.40.0",
    "httpx>=0.28.1",
    "pre-commit>=4.5.1",
    "pyright>=1.1.408",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "pytest-cov>=7.0.0",
    "ruff>=0.14.14",
    "mcp>=1.26.0",
    "allure-pytest>=2.15.3",
]

# Ruff 配置
[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]

[tool.ruff.format]
quote-style = "single"

# Pyright 配置
[tool.pyright]
pythonVersion = "3.12"
typeCheckingMode = "strict"
include = ["src", "tests"]
exclude = ["**/__pycache__", ".venv"]

# Pytest 配置
[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
addopts = ["-v", "--cov=src/agent_core", "--cov=src/agent_app", "--cov-report=term-missing", "--cov-report=xml"]
asyncio_mode = "auto"
markers = ["smoke: 手動 smoke test，需要真實 API（使用 --run-smoke 執行）"]

# Coverage 配置
[tool.coverage.run]
source = ["src/agent_core", "src/agent_app"]
branch = true

[tool.coverage.report]
exclude_lines = ["pragma: no cover", "if TYPE_CHECKING:", "if __name__ == .__main__.:"]
